Added classes Core.cs LogicController.cs Reader.cs Renamed BattleField.cs to GameObject.cs and moved relevant methods to it. put placeholder interface for validator class IValidator.cs Abstracted all logic from main class and left only the core init in the Main() method.
