+++ Batal Field.csproj
@@ -8,11 +8,12 @@
     <ProjectGuid>{35870605-20E7-4668-A5A7-A3926E6D61CE}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>BattleFiled</RootNamespace>
-    <AssemblyName>BattleFiled</AssemblyName>
+    <RootNamespace>BattleField</RootNamespace>
+    <AssemblyName>BattleField</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <TargetFrameworkProfile>Client</TargetFrameworkProfile>
     <FileAlignment>512</FileAlignment>
+    <IsWebBootstrapper>false</IsWebBootstrapper>
     <PublishUrl>publish\</PublishUrl>
     <Install>true</Install>
     <InstallFrom>Disk</InstallFrom>
@@ -25,7 +26,6 @@
     <MapFileExtensions>true</MapFileExtensions>
     <ApplicationRevision>0</ApplicationRevision>
     <ApplicationVersion>1.0.0.%2a</ApplicationVersion>
-    <IsWebBootstrapper>false</IsWebBootstrapper>
     <UseApplicationTrust>false</UseApplicationTrust>
     <BootstrapperEnabled>true</BootstrapperEnabled>
   </PropertyGroup>
@@ -61,16 +61,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BattleField.cs" />
-    <Compile Include="Controllers\MineController.cs" />
-    <Compile Include="Game.cs" />
-    <Compile Include="Interfaces\IFieldObject.cs" />
-    <Compile Include="Interfaces\IFieldObjectController.cs" />
-    <Compile Include="Interfaces\IMineController.cs" />
-    <Compile Include="Models\Mine.cs" />
-    <Compile Include="Models\Point.cs" />
+    <Compile Include="GameObject.cs" />
+    <Compile Include="Core.cs" />
+    <Compile Include="Interfaces\IValidator.cs" />
+    <Compile Include="Utils\LogicController.cs" />
     <Compile Include="Utils\Printer.cs" />
     <Compile Include="Program.cs" />
+    <Compile Include="Utils\Reader.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Properties\" />
