--- Interfaces/IFieldObject.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace BattleFiled.Interfaces
-{
-    interface IFieldObject
-    {
-        int Row { get; }
-        int Col { get; }
-    }
-}
